requirements.txt
autopep8==1.4.4
Django==2.1.8
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pkg-resources==0.0.0
pycodestyle==2.5.0
Pygments==2.5.2
PyMySQL==0.9.3
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
